% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_explorer_plots.R
\name{across_electrodes_f_histogram}
\alias{across_electrodes_f_histogram}
\title{Histogram of F-tests per electrode}
\usage{
across_electrodes_f_histogram(results, ...)
}
\arguments{
\item{results}{results returned by module}

\item{...}{other parameters passed to module output}
}
\description{
Histogram of F-tests per electrode
}
