% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voltage_explorer_plots.R
\name{by_condition_welch}
\alias{by_condition_welch}
\title{Welch periodogram per condition}
\usage{
by_condition_welch(results, ...)
}
\arguments{
\item{results}{results returned by module}

\item{...}{other parameters passed to module output}
}
\description{
Welch periodogram per condition
}
