% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_explorer_plots.R
\name{draw_cut_point}
\alias{draw_cut_point}
\title{Draws an orange, dashed horizontal line at cut. Checks for not null and 
length > 0}
\usage{
draw_cut_point(cut = NULL)
}
\arguments{
\item{cut}{the location(s) of the lines}
}
\value{
the value of cut (invisibly)
}
\description{
Draws an orange, dashed horizontal line at cut. Checks for not null and 
length > 0
}
\examples{
plot(-log10(runif(100)), ylab='p value')
draw_cut_point(-log10(0.05))
}
