% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{draw_img}
\alias{draw_img}
\title{We Are Just Plotting Image(Zmat) Rather Than T(Zmat) As You Might Expect. The Rave_calculators Know This So We Can
save a few transposes along the way}
\usage{
draw_img(zmat, x, y, xlab = "Time (s)", ylab = "Frequency (Hz)", zlim,
  log = "", ...)
}
\description{
The idea here is to to separate the plotting of the heatmap from all the accoutrements that are done in the decorators
}
\author{
John Magnotti
}
