% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_plotting_functions.R
\name{draw_many_heat_maps}
\alias{draw_many_heat_maps}
\title{Draws several heatmaps in a row and (optionally) a color bar}
\usage{
draw_many_heat_maps(hmaps, x, y, xlab = "Time", ylab = "Frequency",
  log_scale = FALSE, show_color_bar = TRUE, DECORATOR, max_zlim = 0,
  useRaster = TRUE, wide = FALSE, ...)
}
\description{
Easy way to make a bunch of heatmaps with consistent look/feel and get a colorbar.
By default it is setup for time/freq, but by swapping labels and decorators you can do anything
}
\author{
John Magnotti
}
