% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_plotting_functions.R
\name{easy_layout}
\alias{easy_layout}
\title{Create a easy layout for multiple plots sharing the same x,y and legend}
\usage{
easy_layout(K, nrows = 1, legend, legend_size = lcm(3),
  legend_side = 4, s_margin = par("mar"), b_margin = par("oma"),
  l_margin)
}
\arguments{
\item{K}{number of plots to be made}

\item{nrows}{number of rows for the plot, default 1}

\item{legend}{expression for generating legend, see "?legend"}

\item{legend_size}{legend width/height, default is lcm(3)}

\item{legend_side}{1 - bottom, 2 - left, 3 - top, 4 - right. Default is 4}

\item{s_margin}{margins within each plots see "?par" for "mar"}

\item{b_margin}{margins for the whole plot see "?par" for "oma"}

\item{l_margin}{legend margin}
}
\description{
Provide easy ways to set plot layouts
}
\author{
Zhengjia Wang, John Magnotti
}
