% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voltage_explorer_plots.R
\name{erp_over_time_plot}
\alias{erp_over_time_plot}
\title{Voltage Time Series Plot}
\usage{
erp_over_time_plot(results, ...)
}
\arguments{
\item{results}{results returned by module}

\item{...}{other parameters passed to module output}
}
\description{
Voltage Time Series Plot
}
