% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_plotting_functions.R
\name{get_palette}
\alias{get_palette}
\title{Function to get builtin color palettes}
\usage{
get_palette(pname, get_palettes = FALSE, get_palette_names = FALSE)
}
\arguments{
\item{pname}{palette name}

\item{get_palettes}{ignored}

\item{get_palette_names}{whether to get palette names}
}
\description{
Function to get builtin color palettes
}
