% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_plotting_functions.R
\name{layout_heat_maps}
\alias{layout_heat_maps}
\title{layout_heat_map}
\usage{
layout_heat_maps(
  k,
  max_col,
  ratio = 4,
  layout_color_bar = TRUE,
  colorbar_cm = 3.5
)
}
\arguments{
\item{colorbar_cm}{The default width chosen (3.5) for the color bar relies on `lcm`. If the function detects the user
is writing to a file (@seealso plotting_to_file), the width is currently forced to 3.0}

\item{k:}{the number of heatmaps, excluding the color bar}

\item{max_col:}{maximum number of columns before creating multiple rows}

\item{ratio:}{heatmap to color bar width ratio (Default 4:1)}

\item{layout_color_bar:}{whether space should be made for the color bar (Default TRUE)}
}
\description{
Create a layout so that heatmaps look nice and you have enough space for the color bar
}
