% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_explorer_plots.R
\name{over_time_plot}
\alias{over_time_plot}
\title{Time Series Plot}
\usage{
over_time_plot(results)
}
\arguments{
\item{results}{results returned by module}

\item{...}{other parameters passed to module output}
}
\description{
Time Series Plot
}
