% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_plotting_functions.R
\name{plot_grouped_data}
\alias{plot_grouped_data}
\title{Description this doesn't do any decoration, it's designed for use with rutabaga::create_frames.
Note that we're using barplot to set the x- and y-range of the plot.
Note Does not handle log axes correctly
param ... extra options to pass to barplot during plot creation}
\usage{
plot_grouped_data(
  mat,
  yvar,
  xvar,
  gvar = NULL,
  types = c("jitter points", "means", "ebar polygons"),
  layout = c("grouped", "overlay"),
  draw0 = TRUE,
  draw0.col = "black",
  ylim = NULL,
  col = NULL,
  ...,
  plot_options = NULL,
  jitter_seed
)
}
\description{
Description this doesn't do any decoration, it's designed for use with rutabaga::create_frames.
Note that we're using barplot to set the x- and y-range of the plot.
Note Does not handle log axes correctly
param ... extra options to pass to barplot during plot creation
}
