% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_plotting_functions.R
\name{pretty_num}
\alias{pretty_num}
\title{Try to guess decimal points of the number and returns string}
\usage{
pretty_num(x, digits = 3, roundup = 5, decimal.mark = ".", ...)
}
\description{
Try to guess decimal points of the number and returns string
}
