% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_explorer_plots.R
\name{windowed_comparison_plot}
\alias{windowed_comparison_plot}
\title{By Trial Plot With Statistics}
\usage{
windowed_comparison_plot(results, ...)
}
\arguments{
\item{results}{results returned by module}

\item{...}{other parameters passed to module output}
}
\description{
By Trial Plot With Statistics
}
