% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa (John Magnotti's conflicted copy
%   2021-08-09).R, R/aaa.R
\name{dev_ravebuiltins}
\alias{dev_ravebuiltins}
\title{Function to load all dev funtions and wrap them within an environment}
\usage{
dev_ravebuiltins(expose_functions = FALSE, reload = TRUE)

dev_ravebuiltins(expose_functions = FALSE, reload = TRUE)
}
\arguments{
\item{expose_functions}{logical indicating whether to expose all dev functions to the global environment}

\item{reload}{logical, do you want to fast-reload the package before load the functions?}
}
\description{
Function to load all dev funtions and wrap them within an environment

Function to load all dev funtions and wrap them within an environment
}
