% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_plotting_functions.R
\name{draw_img}
\alias{draw_img}
\title{RAVE custom image plot}
\usage{
draw_img(zmat, x, y, xlab = "Time (s)", ylab = "Frequency (Hz)", zlim,
  log = "", useRaster = TRUE, ...)
}
\description{
The idea here is to to separate the plotting of the heatmap from all the accoutrements that are done in the decorators. We are just plotting image(zmat) Rather Than t(zmat) as you might expect. The Rave_calculators know this so we can save a few transposes along the way.
}
\author{
John Magnotti
}
