% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_plotting_functions.R
\name{make_image}
\alias{make_image}
\title{RAVE custom image plotter}
\usage{
make_image(
  mat,
  x,
  y,
  zlim,
  col = NULL,
  log = "",
  useRaster = TRUE,
  clip_to_zlim = TRUE,
  add = TRUE
)
}
\arguments{
\item{mat}{z-matrix}

\item{x, y}{z and y axis}

\item{zlim}{value to trim zmat}

\item{col}{vector of colors, color palette}

\item{log}{which axis will be in log scale}

\item{useRaster}{passed to image()}

\item{clip_to_zlim}{whether to clip mat}

\item{add}{logical, whether to overlay current plot to an existing image}
}
\description{
The idea here is to to separate the plotting of the heatmap from all the accoutrements that are done in the decorators.
We are just plotting image(mat) Rather Than t(mat) as you might expect. The Rave_calculators know this so we can save a few transposes along the way.
}
