% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa (John Magnotti's conflicted copy
%   2021-08-09).R, R/aaa.R
\name{rave_title}
\alias{rave_title}
\title{Function make a title for a plot, checks par('bg') to handle dark mode}
\usage{
rave_title(main, cex = rave_cex.main, col, font = 1)

rave_title(main, cex = rave_cex.main, col, font = 1)
}
\arguments{
\item{cex}{the character expansion for the title (default is rave_cex.main)}

\item{font}{the font type (default = 1, plain)}
}
\description{
Function make a title for a plot, checks par('bg') to handle dark mode

Function make a title for a plot, checks par('bg') to handle dark mode
}
\seealso{
title

title
}
